Pull out `buffer` into its own crate
