+++ zed/src/editor/display_map/tab_map.rs
@@ -1,7 +1,7 @@
-use parking_lot::Mutex;
-
 use super::fold_map::{self, FoldEdit, FoldPoint, Snapshot as FoldSnapshot};
-use crate::{editor::rope, settings::HighlightId, util::Bias};
+use crate::util::Bias;
+use buffer::{rope, HighlightId};
+use parking_lot::Mutex;
 use std::{mem, ops::Range};
 
 pub struct TabMap(Mutex<Snapshot>);
